refactor: homeRoute (landingPage & publicProfilePage)
