+++ imports/ui/layouts/BlankLayout.js
@@ -3,7 +3,7 @@ import React from 'react';
 import TopNav from '/imports/ui/common/TopNav';
 
 export default ({content = () => null}) => (
-  <div id="page-top" className="gray-bg">
+  <div>
     <TopNav />
     <div className="wrapper wrapper-content">
       { content() }
